moving App to a folder
